register all models `@pie_core.Model` (#503)
This PR registers all models `@pie_core.Model` instead of
`@PyTorchIEModel` to make them work with
`pie_core.AutoAnnotationPipeline.from_config` in the future.

Additional changes:
 - remove `AutoPyTorchIEModel` (just use `AutoModel` instead)
- re-export `AutoModel` from `pie_core`. NOTE: Previously
`AutoPyTorchIEModel` was exported as `AutoModel`! But if used correctly,
it should not make a difference.
 - register `PyTorchIEPipeline` at name `"pytorch-ie"`
- remove backwards compatibility: `Pipeline` and `AutoPipeline`(use
`PyTorchIEPipeline` in either case)
 - adjust the `README.md`

Note: It is not yet possible to load an "old" pipeline (e.g.
[pie/example-ner-spanclf-conll03](https://huggingface.co/pie/example-ner-spanclf-conll03))
via `AutoAnnotationPipeline` since its config does not has a
`pipeline_type` key. In the future, it should be possible to provide
that as argument in the following manner, but **this still needs a minor
adjustment in `pie_core.Auto.from_config` (TODO: reference issue/PR in
pie-core)**:
```
pipeline = AutoAnnotationPipeline.from_pretrained("pie/example-ner-spanclf-conll03", pipeline_type="pytorch-ie")
```